Turn ARTSR*Log functions into macros
So that the correct file name and line number are used in the log
message.

This introduces a few compiler errors related to format specifiers and
format strings, which we fix here.

+++ Source/SocketRocket/Internal/Security/ARTSRPinningSecurityPolicy.m
@@ -49,7 +49,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (BOOL)evaluateServerTrust:(SecTrustRef)serverTrust forDomain:(NSString *)domain
 {
-    ARTSRDebugLog(self.logger, @"Pinned cert count: %d", self.pinnedCertificates.count);
+    ARTSRDebugLog(self.logger, @"Pinned cert count: %lu", (unsigned long)self.pinnedCertificates.count);
     NSUInteger requiredCertCount = self.pinnedCertificates.count;
 
     NSUInteger validatedCertCount = 0;
